+++ src/.internal/charts/xy/axes/AxisRendererY.ts
@@ -87,6 +87,7 @@ export class AxisRendererY extends AxisRenderer {
 
 		const axis = this.axis;
 		axis.ghostLabel.setPrivate("visible", !this.get("inside"));
+		axis.ghostLabel.set("y", -1000);
 
 		const thumb = this.thumb;
 		const opposite = "opposite";