+++ src/.internal/plugins/exporting/Exporting.ts
@@ -44,12 +44,13 @@ async function _pdfmake(): Promise<any> {
 		import(/* webpackChunkName: "pdfmake" */ "pdfmake/build/pdfmake.js"),
 		import(/* webpackChunkName: "pdfmake" */ "./pdfmake/vfs_fonts")
 	]);
-	let pdfmake = a[0];
-	let vfs_fonts = a[1];
+
+	let pdfmake = a[0].default;
+	let vfs_fonts = a[1].default;
 	const global = <any>window;
 	global.pdfMake = global.pdfMake || {};
-	global.pdfMake.vfs = vfs_fonts.default;
-	pdfmake.vfs = vfs_fonts.default;
+	global.pdfMake.vfs = vfs_fonts;
+	pdfmake.vfs = vfs_fonts;
 	return pdfmake;
 }
 
@@ -120,7 +121,7 @@ export interface IExportingSettings extends IEntitySettings {
 	 * * 1 - fully opaque (default).
 	 *
 	 * NOTE: some image formats like JPEG do not support transparency.
-	 * 
+	 *
 	 * @since 5.2.34
 	 */
 	backgroundOpacity?: number;
@@ -189,7 +190,7 @@ export interface IExportingSettings extends IEntitySettings {
 	 *
 	 * If not set, will use `durationFormat` as set in [[DurationFormatter]] of
 	 * the root element.
-	 * 
+	 *
 	 * @since 5.0.16
 	 */
 	durationFormat?: string,
@@ -199,7 +200,7 @@ export interface IExportingSettings extends IEntitySettings {
 	 *
 	 * If not set, will use `baseUnit` as set in [[DurationFormatter]] of
 	 * the root element.
-	 * 
+	 *
 	 * @since 5.0.16
 	 */
 	durationUnit?: TimeUnit;